!added parse line
